Replace the IValue 'methods' property
We need to distinguish the normal JS methods of IValue and its
descendents from the methods exposed to code running inside the
interpreter.  Previously, this was done by having a 'methods' property
containing the methods to be exposed inside the interpreter.  But that
approach does not support inheritance in the IValue hierarchy.

So instead, use a naming convention for the exposed methods.  For
instance, the 'im_map' JS method is exposed as the 'map' method inside
the interpreter.  This allows inheritance, and also makes it easier to
directly invoke interpreter methods.