Improve JSON encoding to the client to support special types
We want to support types that need to be presented in a special way
in the client (e.g. tables, and lazies when incremental update is
added).  This changes adds that support and makes rendering of JSON to
be sent to the client more consistent.

Types that the client should present specially are encoded as
JSON objects with a '!' property indicating the type.  Because of
this, we need to encode IObject property names that might clash.
This is done by adding an extra '!' char to an property name
consisting of only '!' chars.