Who uses LabelledStatements anyway?
Resolve the ambiguity between an ObjectLiteral and a Block containing
a LabelledStatement in favour of ObjectLiteral.  I'm not sure what
is correct in terms of ECMAscript, but node seems to prefer ObjectLiteral,
and it certainly seems like the preferrable behaviour.

I wouldn't be surprised if my parser change has unintended consequences,
but whatever, I've had enough of this.