Redo result rendering in the client
Previously, the JSON form of the results was fairly directly turned into
DOM nodes.  Now, the JSON gets turned into a tree of Widget objects,
which produce the DOM nodes.  This makes more subtle rendering easier
(we don't rely on generated content).  And becuase Widgets can be kept
around, we can update the displayed results as things change on the server.