Use a mutable context object instead of cont, econt
Previously, we have a lots of references to cont and econt, even
though econt is usually pass through unchanged.  This is already
cumbersome, and extending this approach directly to support return
statements would be intolerable.

So instead of adding another continuation to the argument list of each
procedure, we thought it would be better to bite the bullet and have a
context object that holds the continuations, and makes explicit the
stack-like way that the continuations are used.