Use a mutable context object instead of cont, econt
Before I can implement return statements, there needs to be a way to
skip the remainder of a function body; i.e., to return early. Instead
of adding another continuation to the argument list of each procedure,
we thought it would be better to bite the bullet and have a context
object that maintains the continuation stack as well as error
contexts, and for return, an escape stack.

There's a few details I've papered over for the time being: one is
restoring the error stack when exiting something that introduced a new
error context.

Another is dealing with forcing values while they are already in the
process of being forced: I've assumed this indicates a program error,
rather than implicitly making the forcing expression also lazy.

Another is that some env.evaluate* procedures used to call the error continuation directly, while I trampoline instead.