Incrementally render ILazies to JSON
Previously, rendering to JSON would force ILazies, and only yield the JSON
once forcing completed.  In order to display progress, we need to
be able to return JSON incrementally as ILazy forcing completes.

This commit adds the necessary support to the interpreter, but does not
update the client protocol and client: The server still waits for
JSON rendering to finish before sending it to the client.  But it does
it outside of the interpreter.