Try using multimethods for dispatching widget procedures
There's a fair bit of dispatch-by-type going on; explicit as in
methods assigned to prototypes, and implicit as in instanceof and
if-then-else and tables. Using multiple dispatch helps in a couple of
ways: 1. keeping methods out of prototypes, e.g., out of
Object.prototype in the cases a fallback procedure is desired;
2. being able to specialise on values as well as types.