Don't allow bare JS objects and arrays in interpreter
Previously, the interpreter would coerce JS objects and arrays to
IObjects and IArrays as necessary.  But this was just a hangover from the
pre-IValue days.

This makes it so that objects and arrays are only represented as
IObjects and IArrays.  This makes it easier for IObject to maintain
invariants, such as all values held within an IObject being in their
IValue form.