Clear the ground for industrialised widgeting
By and large this is a rearrangement of the widget code to make it a
bit more uniform and extensible.

'!' values are decoded in the module `types.js`, rather than
implicitly in the widget code. Multiple dispatch is used for providing
extension points (e.g., render). Widgets for displaying output values
now have an intermediate supertype, Presentation, where `widgetize`
lives. Rendering expressions is now done with a widget (and this can
be taken much further of course).