Use widgets pervasively and introduce models
This organises the UI code along observer/observed lines (with some
rough edges). In particular, there's a model for the notebook which is
just a collection of entries that talks to the server to evaluate
expressions when they are input. This is adapted to the (mutually
unaware) view of the notebook.